+++ NEWS.md
@@ -1,3 +1,9 @@
+# beastt 0.0.2
+
+* Add time-to-event capabilities 
+* Fixed bug with the weights of the power priors 
+* Fixed issue with R Version not being high enough for the use of the native pipe
+
 # beastt 0.0.1
 
 * Initial CRAN submission.